requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0

# NetSuite integration
requests-oauthlib==1.3.1

# OpenAI API
openai==1.3.5

# Google Sheets integration
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Snowflake database
snowflake-connector-python==3.6.0

# Document processing
PyMuPDF==1.23.14  # fitz for PDF processing
Pillow==10.1.0  # PIL for image processing
pandas==2.1.4  # Excel file processing
openpyxl==3.1.2  # Excel file support
python-docx==1.1.0  # Word document processing

# Utility libraries
python-dateutil==2.8.2
pytz==2023.3